% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestoR_utils.R
\name{as_tibble.ManifestoCorpus}
\alias{as_tibble.ManifestoCorpus}
\title{\code{\link[dplyr]{as_tibble}} method for \code{\link[manifestoR]{ManifestoCorpus}} class}
\usage{
\method{as_tibble}{ManifestoCorpus}(x)
}
\arguments{
\item{x}{A 'ManifestoCorpus' object.}
}
\value{
A \code{\link[tibble]{tibble}} with
     rows uniquely identifyied by column 'manifesto_id',
     list-column 'data' containing mainfesto text and codes, and
     all other columns recording manifesto meta data.
}
\description{
\code{\link[dplyr]{as_tibble}} method for \code{\link[manifestoR]{ManifestoCorpus}} class
}
