% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto_to_xml.R
\name{combine_head_and_body_}
\alias{combine_head_and_body_}
\title{Combine XML head and body}
\usage{
combine_head_and_body_(head, body)
}
\arguments{
\item{head}{A unit-length XML-formatted string of the data contained in the document head node.}

\item{body}{A unit-length XML-formatted string of the data (chapter, header, sentence, and quasi-sentence nodes)
contained in the document body node.}
}
\value{
A XML-formatted unit-length character vector.
    Out-most tag named 'manifesto', inner tags (in this order) are 'head' and 'body'.
}
\description{
\code{manifesto_df_to_xml} helper
}
