% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_manifesto_df.R
\name{enhance_manifesto_df}
\alias{enhance_manifesto_df}
\title{Enhance manifesto data frame}
\usage{
enhance_manifesto_df(x)
}
\arguments{
\item{x}{A manifesto data frame with the two required columns: 'text' and 'cmp_code'}
}
\value{
The input \code{x} as \code{manifesto.df} object (inherits from \code{\link[tibble]{tibble}}),
     enhanced by column
     'qs_nr' (running quasi-sentence counter),
     'sent_nr' (running sentence counter),
     'role' (indicator, here 'qs' for all rows), and
     'bloc_nr' (enumerates consecutive rows by 'role')

     In addition, the returned \code{manifesto.df} obejct has two attributes:
     \enumerate{
       \item 'annotated': indicates wehtehr or not the input manifesto has been annotated/coded by CMP experts.
       \item 'extra_cols': names of columns added by enhancing the input data frame.
     }
}
\description{
Functon takes a manifesto data frame
     (see \code{as_tibble.ManifestoCorpus} and \code{as_tibble.ManifestoDocument})
     and enhances it with quasi-sentence, sentence, and bloc counters as well as a
     role indicator distinguishing quasi-sentence text (value 'qs'), from title, header and meta text.

     This text-level information is infered from columns 'text' and 'cmp_code'.
}
\note{
As one natrual sentence may contain multiple quasi-sentences, the latter map m:1 to the former.

     For each row, the indicator variable 'role' may assume either of four values:
     \enumerate{
       \item 'qs': quasi-sentence
       \item 'title': the first row(s) with CMP code 'H' or \code{NA} (only in annotated manifestos)
       \item 'header': subsequent rows with CMP code 'H' or \code{NA}  (only in annotated manifestos)
       \item 'meta': in annotated manifestos containing 'H' codes, the row(s) between 'title' and the first 'header' rows
     }
}
\examples{

library(tibble)
library(manifestoEnhanceR)

man <- tribble(
  ~manifesto_id, ~text, ~cmp_code,
  "123", "main title", "H",
  "123", "sub title", "H",
  "123", "Publisher etc", NA_character,
  "123", "first section", "H",
  "123", "This is the first full sentence.", "000",
  "123", "This is the second,", "000",
  "123", "but splitted sentence.", "000",
  "123", "second section", "H"
  "123", "This is the third sentence.", "000"
)

enhanced <- enhance_manifesto_df(man)
class(enhanced)
nrow(man) == nrow(enhanced)
ncol(man) < ncol(enhanced)
attr(enhanced, "annotated")
attr(enhanced, "extra_cols")
}
