% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_manifesto_df.R
\name{enhance_meta_df_}
\alias{enhance_meta_df_}
\title{Enhance meta data (non-quasi-sentence rows) of manifesto data frame}
\usage{
enhance_meta_df_(meta.df)
}
\arguments{
\item{meta.df}{the subset of non-quasi-sentences (headers/titles) rows in a manifesto dataframe}
}
\value{
The input \code{meta.df} \code{\link[tibble]{tibble}} enhanced by column
     'role' (indicator 'title', 'header', or 'meta')
}
\description{
\code{enhance_manifesto_df} helper
}
