% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_manifesto_df.R
\name{enhance_qs_df_}
\alias{enhance_qs_df_}
\title{Enhance quasi-sentence of manifesto data frame}
\usage{
enhance_qs_df_(qs.df)
}
\arguments{
\item{qs.df}{the subset of quasi-sentences (non-header) rows in a manifesto dataframe}
}
\value{
The input \code{qs.df} \code{\link[tibble]{tibble}} enhanced by columns
     'qs_nr' (running quasi-sentence counter) and
     'sent_nr' (running sentence counter)
}
\description{
\code{enhance_manifesto_df} helper
}
