% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_manifesto_df.R
\name{enhance_uncoded_manifesto_}
\alias{enhance_uncoded_manifesto_}
\title{Enhance uncoded manifesto data frame}
\usage{
enhance_uncoded_manifesto_(u.df)
}
\arguments{
\item{u.df}{the 1-row manifesto dataframe recording the entire manifesto text in a unit-length character column 'text'}
}
\value{
The input \code{u.df} \code{\link[tibble]{tibble}} enhanced by column
     'qs_nr' (running quasi-sentence counter),
     'sent_nr' (running sentence counter),
     'role' (indicator, here 'qs' for all rows), and
     'bloc_nr'
}
\description{
\code{enhance_manifesto_df} helper
}
\note{
The function takes the unit-legnth character contained in column 'text' of the 1-row data frame,
     and splits it into sentences.
     Sentences are then numbered, and because the manifesto is uncoded, \code{qs_nr == sent_nr},
     'role' is 'qs' and bloc_nr is 1 for all rows.
}
