% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto_to_xml.R
\name{manifesto_df_to_xml}
\alias{manifesto_df_to_xml}
\title{Convert manifesto to XML}
\usage{
manifesto_df_to_xml(man.df, parse = TRUE, .xml.encode = c("&", "<",
  ">", "#", "\\""))
}
\arguments{
\item{man.df}{A \code{manifesto.df} object (see \code{\link{enhance_manifesto_df}}).}

\item{parse}{logical. If \code{TRUE} (the default), a \code{xml2} \code{xml_document} is returned.
Otherwise, an XML-formatted, unit-length character vector}

\item{.xml.encode}{character vector specifying plain text special characters to be converted to XML-compatible encoded.
Conversion is performed by \code{\link[textutils]{HTMLencode}}.
Defaults to '&', '<', '>', '#' and '"'.
Ignored if \code{NULL}.}
}
\value{
A \code{xml2} \code{xml_document} with parent node 'manifesto',
    or an XML-formatted string if \code{parse = FALSE}.
}
\description{
Function takes \code{manifesto.df} and outputs
    either an XML-formatted string
    or, if \code{parse = TRUE} (the default), a \code{xml2} \code{xml_document}.
}
\examples{
\dontrun{
library(tibble)
library(manifestoEnhanceR)
library(xml2)

man <- tribble(
  ~manifesto_id, ~party, ~text, ~cmp_code,
  "123", "Bloc Party", "main title", "H",
  "123", "Bloc Party", "sub title", "H",
  "123", "Bloc Party", "Publisher etc", NA_character,
  "123", "Bloc Party", "first section", "H",
  "123", "Bloc Party", "This is the first full sentence.", "000",
  "123", "Bloc Party", "This is the second,", "000",
  "123", "Bloc Party", "but splitted sentence.", "000",
  "123", "Bloc Party", "second section", "H",
  "123", "Bloc Party", "This is the third sentence.", "000"
)

enhanced <- enhance_manifesto_df(man)
xml <- manifesto_df_to_xml(enhanced)
class(xml)
xml_attrs(xml)

xml_string <- manifesto_df_to_xml(enhanced, parse = FALSE)
length(xml_string)
cat(xml_string)
}
}
\seealso{
\code{\link{enhance_manifesto_df}}
}
