% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto_to_xml.R
\name{xmlify_chapter_}
\alias{xmlify_chapter_}
\title{XML-ify manifesto chapter}
\usage{
xmlify_chapter_(x, chap.nr)
}
\arguments{
\item{x}{unit-length, XML-fromatted character vector sentenc-quasi-sentence nestings}

\item{chap.nr}{unit-length integer, specifying the chapter number}
}
\value{
A XML-formatted unit-length character vector
    Out-most tag named 'chapter' (w/ attribute 'nr'),
    inner tags are be 'sentence' and 'quasi-sentence' (nested in this order).
    'quasi-sentence' nodes wrap text lines.
}
\description{
\code{manifesto_df_to_xml} helper
}
