% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto_to_xml.R
\name{xmlify_head_}
\alias{xmlify_head_}
\title{XML-ify manifesto meta data}
\usage{
xmlify_head_(x, title = NULL)
}
\arguments{
\item{x}{first row of \code{manifesto.df}}

\item{title}{a character vector specifying the manifesto title
If \code{NULL} (the default) or \code{length(title) == 0},
the title is taken from the input data frame if column exists.}
}
\value{
A XML-formatted unit-length character vector.
    Out-most tag named 'head', inner tags (in this order) may be 'title' and 'p'.
    'p' nodes wrap title lines.
}
\description{
\code{manifesto_df_to_xml} helper
}
