% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto_to_xml.R
\name{xmlify_meta_}
\alias{xmlify_meta_}
\title{XML-ify manifesto meta content bloc}
\usage{
xmlify_meta_(x)
}
\arguments{
\item{x}{rows of \code{manifesto.df} with role 'title', header' or 'meta'}
}
\value{
A XML-formatted unit-length character vector
    Out-most tag named like 'role' of row ('title', header' or 'meta').
    Inner tag is 'p' and contains text lines.
}
\description{
\code{manifesto_df_to_xml} helper
}
